[#26] refactor : DDD
