+++ company/src/main/kotlin/com/sparta/company/infrastructure/client/HubFeignService.kt
@@ -1,6 +1,6 @@
 package com.sparta.company.infrastructure.client
 
-import com.sparta.company.domain.client.HubClient
+import com.sparta.company.application.client.HubService
 import com.sparta.company.presentation.api.response.Response
 import io.github.resilience4j.retry.annotation.Retry
 import org.springframework.cloud.openfeign.FeignClient
@@ -10,9 +10,9 @@ import java.util.*
 
 @FeignClient(
     name = "hub-server",
-    fallbackFactory = HubFeignClientFallbackFactory::class
+    fallbackFactory = HubFeignServiceFallbackFactory::class
 )
-interface HubFeignClient : HubClient {
+interface HubFeignService : HubService {
 
     @Retry(name = "hub")
     @GetMapping("/hubs/company/{hubId}")
