+++ deliveryRoute/src/main/java/com/sparta/deliveryroute/application/service/DeliveryEventService.java
@@ -1,7 +1,7 @@
 package com.sparta.deliveryroute.application.service;
 
-import com.sparta.deliveryroute.domain.client.HubClient;
-import com.sparta.deliveryroute.domain.event.CreateDeliveryEvent;
+import com.sparta.deliveryroute.application.client.HubService;
+import com.sparta.deliveryroute.application.event.CreateDeliveryEvent;
 import com.sparta.deliveryroute.domain.repository.DeliveryRouteRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class DeliveryEventService {
 
-	private final HubClient hubClient;
+	private final HubService hubService;
 	private final DeliveryRouteRepository deliveryRouteRepository;
 
 	@Transactional
