+++ deliveryRoute/src/main/java/com/sparta/deliveryroute/infrastructure/client/HubFeignService.java
@@ -1,7 +1,7 @@
 package com.sparta.deliveryroute.infrastructure.client;
 
-import com.sparta.deliveryroute.domain.client.HubClient;
-import com.sparta.deliveryroute.infrastructure.client.response.RouteResult;
+import com.sparta.deliveryroute.application.client.HubService;
+import com.sparta.deliveryroute.application.dto.response.RouteResult;
 import com.sparta.deliveryroute.presentation.api.response.Response;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,9 +11,9 @@ import java.util.UUID;
 
 @FeignClient(
 		name = "hub-server",
-		fallbackFactory = HubFeignClientFallbackFactory.class
+		fallbackFactory = HubFeignServiceFallbackFactory.class
 )
-public interface HubFeignClient extends HubClient {
+public interface HubFeignService extends HubService {
 
 	@Override
 	@GetMapping("/hubs/routes/id")
