+++ deliveryRoute/src/main/java/com/sparta/deliveryroute/infrastructure/client/HubFeignServiceFallbackFactory.java
@@ -4,9 +4,9 @@ import org.springframework.cloud.openfeign.FallbackFactory;
 import org.springframework.stereotype.Component;
 
 @Component
-public class HubFeignClientFallbackFactory implements FallbackFactory<HubFeignClient> {
+public class HubFeignServiceFallbackFactory implements FallbackFactory<HubFeignService> {
 	@Override
-	public HubFeignClient create(Throwable cause) {
+	public HubFeignService create(Throwable cause) {
 		return null;
 	}
 }
